requirements.txt
fastapi
uvicorn
yt-dlp
openai
anthropic
openai-whisper
python-multipart
aiofiles
